Split off colorizer backend from colorizer
Until now, this comes with no semantical change.
